Percentile Bootstrap Confidence Interval (Wild Bootstrap) - Linear Models Heteroskedasticity

This function calculates confidence intervals for the parameters in heteroskedasticity linear regression models. The intervals are estimated by bootstrap percentile.

Usage

Pboot(model, significance=0.05, double=FALSE, J=NULL, K=NULL,
      distribution="rademacher")

Arguments

model

Any object of class lm;

significance

Significance level of the test. By default, the level of significance is 0.05;

double

If double = TRUE will be calculated intervals bootstrap t and double bootstrap t. The default is double = FALSE;

J

Number of replicas of the first bootstrap;

K

Number of replicas of the second bootstrap;

distribution

Distribution of the random variable with mean zero and variance one. This random variable multiplies the error estimates in the generation of the samples. The argument distribution can be rademacher or normal (standard normal). The default is distribution = rademacher.

Value

A list with the following components:

beta

A numeric vector of length 2 containing the estimated coefficients of the model.

ci_lower_simple

A numeric vector of length 2 containing the lower bounds of the simple bootstrap confidence intervals for the coefficients.

ci_upper_simple

A numeric vector of length 2 containing the upper bounds of the simple bootstrap confidence intervals for the coefficients.

ci_lower_double

A logical vector of length 0 or 2. If `double = FALSE`, this will be a logical vector of length 0. If `double = TRUE`, this will be a numeric vector containing the lower bounds of the double bootstrap confidence intervals for the coefficients.

ci_upper_double

A logical vector of length 0 or 2. If `double = FALSE`, this will be a logical vector of length 0. If `double = TRUE`, this will be a numeric vector containing the upper bounds of the double bootstrap confidence intervals for the coefficients.

J

A numeric value indicating the number of bootstrap resamples used in the simple bootstrap.

K

A numeric value indicating the number of bootstrap resamples used in the double bootstrap, if `double = TRUE`.

References

Booth, J.G. and Hall, P. (1994). Monte Carlo approximation and the iterated bootstrap. Biometrika, 81, 331-340.

Cribari-Neto, F.; Lima, M.G. (2009). Heteroskedasticity-consistent interval estimators. Journal of Statistical Computation and Simulation, 79, 787-803;

Wu, C.F.J. (1986). Jackknife, bootstrap and other resampling methods in regression analysis, 14, 1261-1295;

McCullough, B.D; Vinod, H.D. (1998). Implementing the double bootstrap, 12, 79-95.

Author

Pedro Rafael Diniz Marinho <pedro.rafael.marinho@gmail.com>

See also

Examples

data(schools)
datas = schools[-50,]
y = datas$Expenditure 
x = datas$Income/10000
model = lm(y ~ x)
Pboot(model=model, significance = 0.05, double = FALSE,
      J=1000, K = 100, distribution = "rademacher")
#> Error in eval(predvars, data, env): object 'x' not found